pyproject.toml
[tool.poetry]
name = "currency-converter-api"
version = "0.1.0"
description = ""
authors = ["Ramon Medeiro <r.medeiro10@gmail.com>"]
readme = "README.md"
packages = [{include = "currency_converter_api"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.105.0"
uvicorn = "^0.25.0"
requests = "^2.31.0"
aiohttp = "^3.9.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
